add allProducts api route, set up redux store, create AllProducts and AllProductsRender componenets
