+++ client/store/products.js
@@ -1,0 +1,25 @@
+import axios from 'axios'
+
+const GET_PRODUCTS = 'GET_PRODUCTS'
+
+const defaultProducts = []
+
+const gotProducts = products => ({type: GET_PRODUCTS, products})
+
+export const getProducts = () => async dispatch => {
+  try {
+    const res = await axios.get('/api/products')
+    dispatch(gotProducts(res.data || defaultProducts))
+  } catch (err) {
+    console.error(err)
+  }
+}
+
+export default function(state = defaultProducts, action) {
+  switch (action.type) {
+    case GET_PRODUCTS:
+      return action.products
+    default:
+      return state
+  }
+}
