+++ server/api/products.js
@@ -1,0 +1,12 @@
+const router = require('express').Router()
+const {Product} = require('../db/models')
+module.exports = router
+
+router.get('/', async (req, res, next) => {
+  try {
+    const products = await Product.findAll()
+    res.json(products)
+  } catch (err) {
+    next(err)
+  }
+})
